+++ src/_readable_set_ctx.ts
@@ -3,9 +3,11 @@ import { Readable, readable } from './readable'
 import { get } from './get'
 import type { StartStopNotifier } from './StartStopNotifier'
 import type { Subscriber } from './Subscriber'
-export function _readable_set_ctx<Val extends unknown = unknown>(initial:Val):readable_set_ctx_type<Val> {
+export function _readable_set_ctx<Val extends unknown = unknown>(
+	initial:Val, readable_fn = readable
+):readable_set_ctx_type<Val> {
 	let set = undefined as maybe_undefined<_readable_set_ctx_set_type<Val>>
-	const store = readable<Val>(initial, (in_set=>{
+	const store = readable_fn<Val>(initial, (in_set=>{
 		set = in_set
 	}) as StartStopNotifier<Val>) as Readable<Val>
 	get(store)