+++ src/derived$.ts
@@ -1,0 +1,11 @@
+import { mix_readable_$, Readable$ } from './readable$'
+import { derived, derived_in_fn_type } from './derived'
+import type { Stores } from './Stores'
+export function derived$<Store extends Stores, Val extends unknown = unknown>(
+	stores:Store,
+	in_fn:derived_in_fn_type<Store, Val>,
+	initial_value?:Val
+):Readable$<Val> {
+	const store = derived<Store, Val>(stores, in_fn, initial_value)
+	return mix_readable_$(store)
+}