+++ src/readable.ts
@@ -1,2 +1,2 @@
-export { readable, } from 'svelte/store'
+export { readable } from 'svelte/store'
 export type { Readable } from 'svelte/store'