+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ctx-core/store
 
+## 24.3.0
+
+### Minor Changes
+
+- - readable\$,derived\$,writable\$,Readable\$,Writable\$
+- \_readable_set_a1,\_readable_set_ctx: readable_fn argument
+
 ## 24.2.4
 
 ### Patch Changes